Revised markup to accommodate HTML 2.0 validation
Updated to accomodate the earliest version of HTML standard available on the World Wide Web Consortium Markup Validation Service, HTML 2.0. This version from 1995 does not support the additional image element attributes (`align`, `hspace`, and `width`), so it was changed to a paragraph element to better handle image alignment with the signature text. Removed unnecessary whitespace.