Added Content-Language metadata content
This attempts to address an error related to missing language declaration, a warning about the verbosity of image alternate text with the AIM accessibility verification test. The language issue should be resolved with a language (LANG) attribute on the root HTML document element, but that attribute is no available in HTML 2.0 Strict. Another recommendation of AIM accessibility tool is to use modern struture elements (HEADER, NAV, MAIN, etc.), but those are not avaiable in HTML 2.0 Strict.