Removed Content-Language metadata element
The Content-Language metadata element was added as an attempt to address an error reported by AIM accessibility testing tool called "Document language missing." The recommendation was to use the language attribute (LANG) on an HTML document root element, but the attribute is not available in HTML 2.0 Strict. The Content-Language metadata element was an attempt to resolve the issue using another means, however, it did not satisfy the tool's verification test. In addition, the purpose of the Content-Language metadata element is to direct the browser which language is intended for the audience, not the the language of the document (particularly for multilanguage documents like an document about teaching English to a German viewer).