Added Expires metadata element
The Pingdom speed test tool suggests that Expires  headers should be included in the response for a hyptertext transfer protocol request. The Expires metadata element provides an alternate method of denoting the expiration date and time before the browser should consider it stale and request a refreshed document from the server.